[#24] Fix refresh on click issue
The problem was that the anchors were being triggered (with href to the top of the page) instead of the buttons themselves. Simply removing the anchors and moving the ids to the buttons did the trick. I had put the wrapping anchor elements when first implementing the triggers to make it work but I'm not sure why now and they weren't needed otherwise.