[#74] Add venv activation script
To faciliate the process of activating the venv all the time, this script will simply need to be sourced whenever the need arises and even whenever updates are made to requirements-*.txt or .envrc. Also adds some small version checking of the Python version if it is going to try to install the venv.