[#90] Move connection_test into unit test folder, reinstate simple testRecv.py script
This marks the first real unit test added to this folder (i.e. one that isn't a sample).
The only annoyance with this particular test was the fact that I had to resort to using
delays within the code, which will altogether add 3-4 seconds to the execution time of
running all unit tests within our unit test folder.