[#24] Add feedback to arm page console log for ArmCommandListener
Exception handling added for rover feedback as well as arm feedback.
It was necessary to add command throughput throttling on the client side.
As a result of debugging, extra print statements for debugging have been added to connection class.
Reassingment of ports for receiving/sending:
- Arm: listens for commands on 5005, sends feedback to 5010
- Rover: listens for commands on 5015, sends feedback to 5020

NOTE: The ping command for the Arm MCU does not return the expected 'pong' response,
rather it returns the motor angle messages which it constantly outputs anyway.