[#24][#71] Change rosRoverStart service to be run by odroid instead of root
Additionally add task_handler node to startup nodes in rover.launch.
This will remove the need to make changes to the same files in different
places on the odroid and then have to synch the same files in the cloned repo on the odroid.
Therefore the synchRosRoverStarter script is no longer needed and removed.