[#83] Add get_pid method and implement for edge case with mjpeg streamer
Once again, this is to make sure that the mjpeg streamer process gets properly killed.
Before the overly simple approach was to kill the process id +1 relative to the start_stream.sh process.
Although this seemed consistent behavior at first, it did not end up being consistent so this change was necessary.