[#83] Force resolution down and coerce FPS from 1-5
This greatly reduces the lag of the video stream as well as the time it takes for the system to respond
to a shutdown request while the camera stream is active.
With an average Bandwidth of 5Mbits/sec, the results where about 300-400ms delay in the camera stream and I estimate the latency of the video stream is between 300-700ms.