[#151][#120] Add camera resolution and note in my-webcam.rules
Hard coded the resolution for the cv_camera node launch in 1280x720 (in start_ros_stream.sh)
I chose this resolution because 1080p results in a very laggy video stream. 720p seems to be
both stable and clear for using one camera at a time. Note that when attempting to save an image
with the 720p resolution set, the actual image resolution is 800x600 (which meets ERC's requirements)
It is possible this happens due to the resolution not vibing the webcam.

Added a note to my-webcams.rules regarding ID_PATHs. When swapping between webcams with the same USB port,
it is possible that the ID_PATH can change. This happened when swapping between two different types of webcams.