[#86] Add basic feedback for current index position of turntable
The value is displayed in a simple readonly input field.
Ideally this would actually trigger a redraw of the wheel.
For now the purpose is so that if the wheel is out of synch the user
can keep turning the turntable until they see the index at 0,
then refresh the page so that the wheel is correctly drawn.