[#484] Hack till MVP finish line for SAR filming
This certainly is not a perfect solution, but hey, it works on my PC ¯\_(ツ)_/¯

Before merging to master I would definitely recommend testing it with rover cameras (IIRC, logitech c920s), and adjust the values for the mapping function for the start/end points of squares accordingly..

Long term it might be simpler (and better) to use opencv for not only drawing but for detecting AR tags. If we replace ar_track_alvar ros package with opencv then we have access to convenience functions that make drawing contours around detected objects a two-liner, rather than this DIY spaghetti solution we have done here.